% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNumeralSystem.R
\name{convertBin2Dec}
\alias{convertBin2Dec}
\title{Convert binary value to decimal value}
\usage{
convertBin2Dec(x)
}
\arguments{
\item{x}{Binary vaulues (numeric or character type)}
}
\description{
Convert binary value to decimal value
}
