% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNumeralSystem.R
\name{convertDec2Bin}
\alias{convertDec2Bin}
\title{Convert decimal value to binary value}
\usage{
convertDec2Bin(x)
}
\arguments{
\item{x}{Decimal vaulues (numeric or character type)}
}
\description{
Convert decimal value to binary value
}

