% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNumeralSystem.R
\name{convertDec2Hex}
\alias{convertDec2Hex}
\title{Convert decimal value to hexa value}
\usage{
convertDec2Hex(x)
}
\arguments{
\item{x}{Decimal vaulues (numeric or character type)}
}
\description{
Convert decimal value to hexa value
}

