% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNumeralSystem.R
\name{convertHex2Dec}
\alias{convertHex2Dec}
\title{Convert hexa value to decimal value}
\usage{
convertHex2Dec(x)
}
\arguments{
\item{x}{Hexa vaulues (character type)}
}
\description{
Convert hexa value to decimal value
}
