% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFunctionInputFromTest.R
\name{extractFunctionInputFromTest}
\alias{extractFunctionInputFromTest}
\title{Extract function input while running test code}
\usage{
extractFunctionInputFromTest(functionName, testPkg = ".", baseTestCode, cache)
}
\arguments{
\item{functionName}{A string.}

\item{testPkg}{Same as \code{pkg} in \code{\link[devtools]{test}}. Package description, can be path or package name. See \code{\link[devtools]{as.package}} for more information}

\item{baseTestCode}{Same as \code{filter} in \code{\link[devtools]{test}}. Only tests with file names matching this regular expression will be executed. Matching will take on the file name after it has been stripped of "test-" and ".R"}

\item{cache}{Directory in which to store exported function input}
}
\description{
Extract function input while running test code
}

