% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStatusOfProducingTestAnswer}
\alias{getStatusOfProducingTestAnswer}
\title{Get the status of producing test answer}
\usage{
getStatusOfProducingTestAnswer(returnStatus = TRUE)
}
\arguments{
\item{returnStats}{A logical controls return status or not}
}
\description{
Get the global option \code{produceTestAnswer}
}
\seealso{
\code{\link{startProducingTestAnswer}}, \code{\link{stopProducingTestAnswer}}
}

