% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grepMulti}
\alias{grepMulti}
\title{Multiple pattern matching.}
\usage{
grepMulti(pattern, x)
}
\arguments{
\item{pattern}{a character string to be matched in the given character vector.}

\item{x}{a character vector to be matched to.}
}
\value{
\code{grepMulti} returns a vector of the indices of \code{x} that
matched with the elements of \code{pattern}.
}
\description{
\code{grepMulti} matches \code{pattern}, in this case a user name(s)/appliance(s),
to a character vector \code{x} of all users/appliances, respectively.
}
\examples{
user.name <- c('David')
appliance.name <- c('Fridge')
user.list <- c('Kanye', 'David', 'John')
appliance.list <- c('AirWasher', 'Fridge', 'MicroOven', 'Fridge')

grepMulti(user.name, user.list) # 2
grepMulti(appliance.name, appliance.list) # 2 4

}
\seealso{
\code{\link{grepMetaData}}: searches meta data given a targeted user and/or
appliance and returns the selected row of data. \cr
\code{\link{searchMetaData}}
}
