% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowMedian}
\alias{rowMedian}
\title{Form Row Medians}
\usage{
rowMedian(x)
}
\arguments{
\item{x}{an array of two dimensions,
containing numeric, integer values, or a numeric data frame.

Form row medians for numeric arrays (or data frames).}
}
\description{
Form Row Medians
}
