% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicalOperator.R
\name{setdiffNumericForced}
\alias{setdiffNumericForced}
\title{Set difference of converted inputs.}
\usage{
setdiffNumericForced(x, y)
}
\arguments{
\item{x}{first set of characters.}

\item{y}{second set of characters. Note that the function is order dependent,
and this is addressed in \code{\link{setdiffNumericForcedCommutative}}.}
}
\description{
\code{setdiffNumericForced} converts numeral character inputs of two sets
into numeric and finds the difference between the sets.
}
\examples{
first <- c('3', '4', '5')
second <- c('1', '3', '6')

setdiffNumericForced(first, second)
setdiffNumericForced(second, first)

}
\seealso{
\code{\link{setdiffNumericForcedCommutative}}: takes the union of the
commutative \code{setdiffNumericForced}, returning the total set difference. \cr
\code{\link{searchMetaData}}
}

