% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicalOperator.R
\name{setdiffNumericForcedCommutative}
\alias{setdiffNumericForcedCommutative}
\title{Order-independent difference of a set of converted numerics.}
\usage{
setdiffNumericForcedCommutative(x, y)
}
\arguments{
\item{x}{first set of differences.}

\item{y}{second set of differences.}
}
\description{
\code{setdiffNumericForcedCommutative} takes the union of values returned by
\code{setdiffNumericForced(x, y)} and \code{setdiffNumericForced(y, x)},
returning the total set of unlike integers between \code{x} and \code{y}.
}
\examples{
# Here, a and b can be thought of as results from setdiffNumericForced(x, y)
# setdiffNumericForced(y, x), respectively.
a <- c('4', '5')
b <- c('1', '6')

setdiffNumericForcedCommutative(a, b)

}
\seealso{
\code{\link{setdiffNumericForced}}: converts character inputs of two sets
into numeric and finds the set difference. \cr
\code{\link{searchMetaData}}
}

