% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{startProducingTestAnswer}
\alias{startProducingTestAnswer}
\title{Start producing test answer}
\usage{
startProducingTestAnswer()
}
\description{
Set the global option \code{produceTestAnswer} as \code{TRUE}
}
\seealso{
\code{\link{stopProducingTestAnswer}}, \code{\link{getStatusOfProducingTestAnswer}}
}

