% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopProducingTestAnswer}
\alias{stopProducingTestAnswer}
\title{Stop producing test answer}
\usage{
stopProducingTestAnswer()
}
\description{
Set the global option \code{produceTestAnswer} as \code{FALSE}
}
\seealso{
\code{\link{startProducingTestAnswer}}, \code{\link{getStatusOfProducingTestAnswer}}
}
