% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorizedMedian}
\alias{vectorizedMedian}
\title{Vecorized median}
\usage{
vectorizedMedian(x, chunkSize)
}
\arguments{
\item{x}{A numeric vector}

\item{chunkSize}{An integer}
}
\description{
Calculate median for each chunks of given vector with customizable chunk size.
}

